Bug #21348684 SIGABRT DURING RESIZING THE INNODB BUFFER POOL ONLINE WITH MEMORY FULL CONDITION
Problem:

InnoDB buffer pool resize is done by background thread.
Consider the following scenario:

1.  User issues "set global innodb_buffer_pool_resize = X;", where
    X is very high and allocation of such large memory is bound to fail.
2.  The system variable innodb_buffer_pool_resize is immediately
    set to X, and then an event is generated to resize buffer pool.
3.  The background thread wakes up with the event generated.  It
    proceeds to do buffer pool resize.
4.  The background thread may fully or partially fail depending on the
    available memory in the machine.

The memory allocation failure was considered as a fatal error.  Also,
the system variable was not updated to the actual value of the innodb
buffer pool.

Solution:

If memory allocation fails, it should not be considered as a fatal failure.
And update the system variable to the actual value of the innodb buffer
pool after doing resize.

rb#9581 approved by Deb.