Bug#21512842: Sig 11 in check_column_grant_in_table_ref
For a specific field name resolution pattern (a GROUP BY expression
with a column that matches an aliased column), the cached_table
field was not assigned correctly. This field was later required when
resolving the inner query as a derived table, which caused a NULL
pointer dereference.

Fix: Assign the field.