Bug #21455603 	REGRESSION: REPLICATION FAILS FOR INSUFFICIENT PRIVILEGES
Refactoring of SHOW handler in WL6629 left some flaws in privileges that
were exposed by this bug report. When @@global.show_compatibility_56 is OFF,
access to show variables is made through performance_schema so that
a slave account with minimal privileges could not actually connected to the master
anymore.
That happens to the regular slave as well as to the semi-sync slave.

Fixed with converting SHOW-VAR-LIKE queries into SELECT-based ones
working on target variables directly.