BUG#20481175 - ALTER ADD FTS INDEX ASSERT TRX->DICT_OPERATION_LOCK_MODE == 0 || TRX->DICT_OP...
This bug is mainly due to when a trx is selected as a victim for some other
trx with higher priority and the rollback is undergoing, the trx should not
be used in other threads until it's rolled back completely. Because the
rolling back trx would require the RW_S_LATCH for dict_operation_lock during
rollback, if the trx is also running by other threads, the failing assertion
could be triggered. In current design, the thread with higher priority who
wants to kill the victim trx would enter InnoDB first, and any follow up
entrance would not be checked. That's why the trx can enter InnoDB again in
some other thread.

The fix should be simple, when one thread wants to enter InnoDB, it should
also check if there is an asynchronous rollback for this trx, and the thread
itself is not the killer thread. If so, it have to wait for the asynchronous
rollback.

Reviewed-by: Sunny Bains <sunny.bains@oracle.com>
RB: 9780