Bug#21360188 	BACKUP CRASHES DATA NODE DUE TO FAILED NDBREQUIRE AT LINE 1184
The LCP Fragment Scan watchdog checks that LCP related fragment scans make a
reasonable amount of progress.  If progress is insufficient then the node
is shutdown.  As part of the shutdown, some debugging info it produced.

The producton of the debugging info was itself buggy and took a testing oriented
path which had strict assertions which are not always met in real systems.  This
caused the node to fail with an internal error, rather than an LCP Fragment
watchdog related shutdown.

This patch removes calls through the over-strict assertions, and also removes
duplicated debugging output.

Tested with testNodeRestart -nLCPScanFragWatchdog T1 + a concurrent backup.