Bug#21139722: Assertion failed: !(used_tables() & ((table_map) 1) << size...))
This problem is hit when merging a derived table into an outer query
block. When doing Item::resolved_used_tables(), a scalar subquery is
visited (the left-hand side of the inner-most IN subquery).
This function asserts when meeting an outer reference, but an outer
reference is actually valid here.

The problem is that Item::resolved_used_tables() is insufficient:
1. It is not properly implemented for all kinds of Item classes, and
2. It lacks a clear definition of what it is supposed to do.

This is examplified by the default implementation for class Item, which
evaluates a rather arbitrary assert.

The first problem occurs because Item::resolved_used_tables() is
implemented only for a few classes. To make it reach all items, we
either have to make implementations for expression nodes and condition
nodes, etc. Or we can replace it with a walk() function. This is the
chosen approach, and the new function used for walk() is named
used_tables_for_level(). See below for explanation of the name.

Another example demonstrating insufficient implementation is this case:

  CREATE TABLE t1(a INTEGER);
  SELECT (SELECT a FROM t1 AS t2
          WHERE a IN (SELECT t1.a+t2.a FROM t1 AS t3))
  FROM t1 AS t1;

When the IN subquery is transformed into a semi-join, the used tables
information of the selected expression t1.a+t2.a is recalculated with
help of ::resolved_used_tables(). We see that before transformation,
this expression is an outer reference (pointing to the query block at
level 0 and 1). After transformation, t1.a is still an outer reference,
but t2.a has become a local reference. So used tables information for
the expression should be OUTER_REF_TABLE_BIT | 1.

Thus, we can better formulate the requirements for this function:

  Recalculate used tables information for an expression for
  a specific query block.

The expression t1.a+t2.a is moved from the eliminated query block into
the parent query block, so it is recalculated for the parent query block.

The transformation also affects the scalar subquery, so its used tables
information is recalculated for the query block where it is located,
which is the outer-most query block.

Since the transformation occurs in a strict ascending order of query
blocks, we are concerned about the query block at a certain level,
and hence the new function is named used_tables_for_level().

A specific implementation is only needed for Item_field. If the column
is resolved from the same query block as the requested one, the specific
table number is returned. If the column is resolved from a query block
that is outer to the requested one, it is an outer reference.