Bug#21139402 ASSERTION FAILED: LENGTH > 0 && KEYPARTS != 0, CRASH IN JOIN::OPTIMIZE_KEYUSE
select 1 from t1 where 1 in
(select (t2.c is null) in (select t1.a from t1 where t1.b)  from t2);

Scenario, in time order:

1) FIXING 'c IS NULL':
Item_func_isnull::fix_fields():
c is a pk column, cannot be NULL, so this item is marked as constant
with a value of FALSE:
  virtual void update_used_tables()
  {
    if (!args[0]->maybe_null)
    {
      used_tables_cache= 0;            /* is always false */
      const_item_cache= 1;
      cached_value= (longlong) 0;
    }

2) DOING IN->EXISTS FOR SELECT#3:
Item_in_subselect::single_value_transformer():
"c IS NULL" wrapped into Item_ref which is is injected into WHERE:
"WHERE b AND (c IS NULL)=t1.a"
Item_ref is NOT marked as an outer reference because 'c IS NULL' is
constant:

    // Make the left expression "outer" relative to the subquery
    if (!left_expr->const_item())
      left->depended_from= select->outer_select();

(this if() is introduced by:
Bug#16369522: Assertion failed: !tables ||
thd->lex->is_query_tables_locked() with comment:
"   Notice also that const items need not be designated as outer.
    Ignoring those occurrences was required to preserve multiple equality
    handling for equalities involving constant values.";
it might be why this bug is a regression)

Same function does fix_fields() on this new WHERE, which does nothing
as 'c is NULL' is already fixed.

3) SELECT#2 IS SEMIJOIN-MERGED INTO SELECT#1

We do fix_after_pullout on sj condition and thus on SELECT list of select#2
and thus on select#3 and thus on WHERE of select#3 and thus on
the previously injected 'c IS NULL'; this is fix_after_pullout for 'c
IS NULL':
  if (arg_count)
  {
    for (arg=args, arg_end=args+arg_count; arg != arg_end ; arg++)
    {
      Item *const item= *arg;
      item->fix_after_pullout(parent_select, removed_select);
      used_tables_cache|=     item->used_tables();
      not_null_tables_cache|= item->not_null_tables();
      const_item_cache&=      item->const_item();

'item' is t2.c. The last lines above declare that "IS NULL uses
t2" and is not constant (anymore!).
So we now have a non-constant item in select#3, not declared as outer
reference, involving a table from select#1 (t2 is in select#1 now). This is
wrong.

Later we have a keyuse for select#3 for '(c IS NULL)=a',
so keyuse->used_tables= used_tables of 'c is NULL' = map of t2, wrong,
as t2 is not in select#3. So we try to access a join_tab for t2 in
select#3 and crash.

Fix:
when we pull out an Item_func, if it's already constant there's no
reason it could become non-const; and if it's constant it doesn't
really use tables, so there's no reason to recalculate
used_tables_cache and friends.