Bug#21491442 variant::forced_return() [with t = json_scalar*]: assertion `false' failed.
This problem here is another issue of the problem with state in
Items. Normally, at the end of a statement execution, the items are
cleaned up. However, code for generated columns is not presently
cleaned up.

This patch adds cleanup for cg code when we close the tables for the
statement (close_thread_tables), and re-fixes the gc code to be ready
for new execution (since cleanup clears the fixed flag) in
TABLE::init.  This is necessary since in some cases expression tree
objects are used to cache execution information, e.g.
Arg_comparator::json_scalar.

Added a new test case for the repro for all combinations:

      - normal and prepared statements
      - normal and temporary tables