Bug #21039264: HANDLE_FATAL_SIGNAL (SIG=11) IN MI_EXTRA |                STORAGE/MYISAM/MI_EXTRA.C:44
Issue:
-----
The updatable property of a view is set during creation of
the view. If the underlying table is dropped and re-created
as a non-updatable one, the original view's updatable
property is not revised accordingly. This will cause a
problem when an attempt to insert or replace into the view
is made. This problem is specific to views with multiple
tables/views. It also does not occur with update statements.

SOLUTION:
---------
This problem affects only the insert/replace statements
because only the updatable property of the top view is
tested. The problem does not occur for update statements,
since the updatable property of each underlying table/view
is tested.

For a view with a single table, the updatable property is
set based on the underlying table's property. The same
should be done for views with multiple tables/views.

This fix is specific to 5.6. The problem was fixed in 5.7+
by WL#5275 and Bug#20515155. But this is not a backport
since the code has changed significantly.