Bug #21462732 MYSQLPUMP --DEFER-TABLE-INDEXES MISHANDLES IN SPECIFIC CONDITION
With --defer-table-indexes option in mysqlpump, mysqlpump would parse the
SHOW CREATE TABLE output and extract indexes and constraints so that they are applied
on table after rows are dumped. This parsing of the table DDL had a problem because of
which there was a syntax error during restore of dump file. This patch fixed this
parsing issue.