Simplify the extension of the lock free hash
Serialize the code that copies entries from one array to the next one(s)
and then frees the unused array. This has a performance impact only on a
workload that does an aggressive grow (uncommon), but is much simpler
and thus easier to maintain.

WL#7170 InnoDB buffer estimates for tables and indexes