Bug#21552524: GCC CRASHES WHEN COMPILING RELEASE BUILD ON ARM
GCC crashes when compiling server InnoDB source files as part
of release build on ARM64. This is due to a GCC bug that is
present in at least GCC 4.9.2 and GCC 5.1.1.

This patch adds a workaround - compile the affected files with -O0
(on ARM64 only). This makes it possible to build release builds
on ARM64 while waiting for a fixed GCC version.

Note that the GCC bug does not seem to affect ARM32.