Bug#21574096: ERROR LOG FILE DISABLED IF MYSQLD IS               STARTED AS A WINDOWS SERVICE
The problem was that the error log file by mistake was disabled
if mysqld was started as a Windows service. This was a regression
introduced by the patch for Bug#21328041.

This patch fixes the problem by enabling the error log file on
Windows unless --help or --console is used (i.e. regardless of
if mysqld is started as a service or not).

Note that this bug did not affect messages being written to
the Windows event log.