Bug#21055139 SUBQUERY HAVING COUNT WITH GROUP BY GIVES INCORRECT RESULT
For 'x in (SELECT y)', subquery materialization is used.
x and y are here a list of expressions.

The result of 'select y' is stored into a temporary table; duplicates
are eliminated:
1) by default this is done by creating a unique index over the columns
of the tmp table
2) but here the length of "y" is so big that it's not possible; a
hidden hash column, indexed, is added; because of possible collisions
(two different "y" having the same hash value), the hash index is not
unique; if, upon insertion, there is a collision with an existing hash
value, the values of "y" are compared to decide if this is a duplicate
or not (see check_unique_constraint() in
Query_result_union::send_data()).
Possibility (2) is new in 5.7, added by WL#6711 "Support InnoDB as
additional storage engine for tmp table".

Then we look "x" up in the tmp table's index (ha_index_read_map()).
In case (1) above, the index is unique, so a single lookup is enough.
In case (2), the index is not unique, so we can have:
hash y
1    A
2    B
and if we are searching for "x=B", the first lookup is a false positive.
Thus we must do all lookups until we find a match (this matching is
ensured by the condition "x=y" in
"subselect_indexsubquery_engine::cond" created in
subselect_hash_sj_engine::setup()).
This problem was missed in WL#6711.

Fix: if using hash column, index is not unique, so set unique=false.
No new testcase is added, as an existing testcase in
subquery_sj_mat_nosj.test is covering the case (transitions from bad
to good result). Notice how it has ("time","time") as value for "y";
this gives a hash value of 0; so does ("lpjdzvkp","lpjdzvkp"); so does
any pair of two identical strings (CRC of second string is XORed with
CRC of first steing). So we have different "y" for a same hash value.
Changes in EXPLAIN: are about subqueries using hash_field; they must
try multiple lookups.