Fixed ATR testcase 'test_event .. -n Apiv2HQE-latestGCI'
An event subscription was effectively executed twice
(and unsubscribed only once)

The testcase called the local createEventOperation()
which as part of it work also called ::execute() on the
subscription.

Then the callee also executed the returned event operation
such that another subscription for the same event was created.
This effectively created dangling, non-dropped subscriptions
when the test terminated.

Note: it is questionable whether such double subscribe should
be allowed at all by the event-API. I would really have expected
this to give an 'already subscribed'-error.