Bug #18979515: HANG/INFINITE LOOP IN                ITEM_SUM::CLEAN_UP_AFTER_REMOVAL
Issue:
-----
Item_sum maintains a circular chain of registered objects
that keeps track of all the aggregate functions used in a
query. Item_sum::clean_up_after_removal will remove the
Item_sum object from this list. The problem occurs when the
same Item_sum object is revisited (through an Item_ref in
the HAVING clause inserted by the IN to EXISTS
transformation). On the second visit, the item's next
pointer still points to an element in the list, although
the item itself is no longer in the list. The loop will
start traversing the list, but it will never detect the
end since the start element is no longer in the list, so it
loops forever.

SOLUTION:
---------
By assigning Item_sum object's next pointer to NULL, we
make sure that a revisit to clean_up_after_removal will not
result in a loop.