Bug#21296553: WARNINGS SENT TO STDERR/OUT EVEN WHEN LOGGING SET TO SYSLOG
The problem was that some server error messages during server boot are
written to stderr/out even if syslog and/or error log file is enabled.
This happened because proper error logging is only enabled after the
relevant options have been parsed and the server knows where to write
the messages. This is especally a problem if mysqld has been detached
so that stderr/out messages are simply lost.

This patch fixes the problem by buffering error log messages until
the relevant options have been parsed. The existing Buffered_logs
implementation, which was only partially in use, has been removed
in favour of a similar implementation hidden behind the standard
error log API. This means that the error log API can be used at
all times.

Note that if the server terminates during startup before the
error file log has been opened (we know which file to open),
error messages will be written to stderr as before.