Refactoring: Move ACC_TO_REQ into ACCKEYREQ.
The scan lock take over (ACC_TO_REQ) can be seen as a separate
(non-abortable) operation, but is always used in combination with
a primary key operation (ACCKEYREQ).

Moving the call to ACC_TO_REQ into ACCKEYREQ will make it possible
to add verifications that lock taken over is for the same element
as the primary key operation.

During node stops and starts the scan lock taken over may be the
wrong one.  The scan locks are only identified within a fragment
with a scan number and an operation index.  These numbers are leap
numbers and on node restart starts from zero again and are likely
identical to some scan lock from before node stop/start if fragment
have switch primary replica.