Bug #19985318: DATE_FORMAT: STACK BUFFER OVERFLOW IN                INT10_TO_STR
Issue:
-----
When the number of days calculated by the DATE_FORMAT
function is negative, there is incorrect behavior.

SOLUTION:
---------
DATE_FORMAT function does not consider a negative value
for the number of days as possible. Hence supplies a
positive radix value to the INT10STR function. This results
in misinterpretation of the actual value of the days. A
negative radix value should be supplied when the number of
days is negative.