Bug#21296553: WARNINGS SENT TO STDERR/OUT EVEN WHEN LOGGING SET TO SYSLOG
Post-push fix: Fix sporadic failure in main.help_verbose.

The problem was that this test asserts that no warnings or
errors are written to the error log. After the patch for
Bug#21296553 warnings/errors that occur during early server
startup are now written to the error log. This caused the
test to fail on platforms where e.g. max number of connections
where adjusted due to OS configuration.

This patch fixes the problem by rewriting the tests so it
only checks for errors in the error log. Warnings like
the one mentioned above are normal and expected, and should
not be silenced as they affect output of --help --verbose.