Fix crashing ATR Multi* test
Both the test_event testcases Multi & Multi_NR called
the initilaizer createAllShadows which filled the global
Vector pShadowTabs.

As this Vector was not cleared before it was filled, the
second call to createAllShadows append more 'tabs' on top
of it previous contents.

That caused a crash in Multi_NR when it ran after the
Multi test case