Bug#21616810 CONNECTION HANG OR ASSERTION 0 IN THD::SEND_STATEMENT_STATUS, SQL_CLASS.CC
Problem: If the second or third argument of ST_AsGeoJSON() is a NULL
value, the function returns 'error_json()' without calling my_error().
This causes an assertion to be raised in debug mode for some queries.

Fix: Return 'false' instead of 'error_json()' if one of the
arguments is a NULL value.