Bug#21541481 MEMORY LEAK OF ALLOCATIONS MADE IN VAL_JSON_FUNC_FIELD_SUBSELECT
When comparing rows, the array Arg_comparator::comparators. In
Arg_comparator::cleanup, we called delete_json_scalar_holder on the
immediately held Json_scalar_holder, but not on any such inside the
array used for comparing rows, hence the memory leak. So, added that
and the repro test case to json.test.

Note: Arg_comparator's "a" argument (whose "cols" method we could
potentially use to determine the number of members in the
"comparators" field) could already have been destroyed at the point of
execution of Arg_comparator::cleanup.  Therefore we store the number
separately instead.