Patch for bug#21651536:
POLLEVENTS() RETURNS INCORRECT 'LATESTGCI'

The internal NdbEventBuffer::m_latestGCI was not always updated
when a completed epoch of event buffers is inserted into the
event queue. Thus later calls to pollEvents() fails to get and return
a correct GCI of the events available in the buffers.

Added a testcase and asserts catching if nextEvent() is about
to return an event with a higher GCI than previously polled,