Bug#21464787 THE CHANNEL SERVICE INTERFACE LACKS A METHOD TO CHECK THE RECEIVER THREAD ID.
Currently the channel service interface lacks a method to check the
receiver thread id while having this method for the applier thread. In
fact the interface should be generic and always allow to execute the
same methods on both threads whenever it makes sense to.

The method is now generic and allows to check either the receiver or
the applier thread. It shall be useful to identify if the thread that
executed some hook is the thread from channel A or B.