WL#8756  Deprecate binlog_max_flush_queue_time in 5.7
The variable binlog_max_flush_queue_time was introduced in 5.6
in order to fine-tune the Binlog Group Commit. Due to changes
in 5.7, the option has no effect any more. Therefore we
deprecate the variable in 5.7 and plan to remove it in 5.8.

This worklog will generate a deprecation warning in 5.7 and 5.8.
It will not remove the variable in 5.8.