BUG#21266784 - PATCH FOR 21052754 SHOWS UP TO AN 11% DROP IN PERFORMANCE FOR UPDATE OPS
The patch would fix the L1 cache miss issue. The on/off check and counters
per mutex should now be on the same cache line.

The patch also fix an issue that we should distinguish BlockWaitMutex from
WaitMutex.

Finally, the patch does lots of code cleanups to improve the performance,
by doing statstics in batch mode.

The regression observed is 4-5% now.

The original patch is provided by Sunny.

Reviewed-by: Jimmy Yang <jimmy.yang@oracle.com>
RB: 9608