Bug#20146455: FIND_KEY_CI RETURNS NULL, CAUSES CRASH IN FILL_ALTER_INPLACE_INFO
The problem was that the server could terminate if a generated
foreign key index was renamed by the same ALTER TABLE statement
that added a new foreign key with the same name.

The reason this happened is the server figured out that the
two foreign keys would result in two generated keys so similar
that one of them was redundant. If the redundant key was the
same key that was renamed, the code which handles index renames
in fill_alter_inplace_info() could not find the key to rename.

This patch fixes the problem by treating a renamed key as if
it is no longer generated.