Fixed bug#21664573: GC: WRONG ALTER TABLE FLAGS SET BY PARSER
It's incorrect to use Alter_inplace_info flags for syntax of generated column.
Alter_inplace_flags are part of the handler API and are used by the DDL code
to communicate to the SE which operations to execute on.  Alter_info flags are
used by the parser to communicate to the DDL code which operations where
specified by the user.

This patch replaces Alter_inplace_info flags with Alter_info flags.