BUG#21641780 - CRASH ON 'START/STOP SLAVE FOR CHANNEL 'GROUP_REPLICATION_APPLIER'
Problem :

When we have a group running a user can execute:

START SLAVE FOR CHANNEL 'group_replication_applier';
STOP SLAVE FOR CHANNEL 'group_replication_applier';

The STOP SLAVE command sets the FDE (FormatDescriptionEvent) of the
channel to NULL. So an attempt to queue any event after that causes a
crash in the Transaction_boundary_parser::get_event_boundary_type as a
NULL value is being used.

For the normal slave, queue_event() is only executed by the IO thread
that controls the removal of the FDE so there is no issue there. In
Group Replication we set our a "local" master info object (that is
replaced if the channel IO thread connects) for scenarios where there is
no IO thread and that gets deleted due to STOP SLAVE.

Solution :

The solution implemented here are :

Inside rpl_slave.cc::queue_event() check if FDE is NULL, if it is
then we throw an error.
Disallow execution of START/STOP SLAVE and IO_THREAD on FOR CHANNEL
'group_replication_applier';
We also disallow the execution of SHOW SLAVE STATUS for channel
group_replication_applier'.